:construction: WIP :recycle: Move `ResourcePool` init from functions to methods
