+++ CPAC/pipeline/engine/engine.py
@@ -14,36 +14,24 @@
 
 # You should have received a copy of the GNU Lesser General Public
 # License along with C-PAC. If not, see <https://www.gnu.org/licenses/>.
+"""C-PAC pipeline engine."""
+
 import ast
 import copy
 import hashlib
 import json
 import logging
 import os
-import warnings
 
 from nipype import config
 
 from CPAC.pipeline import nipype_pipeline_engine as pe
 from CPAC.pipeline.nodeblock import NodeBlockFunction
-from CPAC.resources.templates.lookup_table import lookup_identifier
-from CPAC.utils.datasource import (
-    create_anat_datasource,
-    create_func_datasource,
-    create_general_datasource,
-    ingress_func_metadata,
-    resolve_resolution,
-)
-from CPAC.utils.interfaces.function import Function
 from CPAC.utils.monitoring import (
     getLogger,
     LOGTAIL,
     WARNING_FREESURFER_OFF_WITH_DATA,
     WFLOGGER,
-)
-from CPAC.utils.utils import (
-    read_json,
-    write_output_json,
 )
 
 
@@ -537,708 +525,7 @@ def wrap_block(node_blocks, interface, wf, cfg, strat_pool, pipe_num, opt):
     return (wf, strat_pool)
 
 
-def ingress_raw_anat_data(wf, rpool, cfg, data_paths, unique_id, part_id, ses_id):
-    if "anat" not in data_paths:
-        WFLOGGER.warning("No anatomical data present.")
-        return rpool
-
-    if "creds_path" not in data_paths:
-        data_paths["creds_path"] = None
-
-    anat_flow = create_anat_datasource(f"anat_T1w_gather_{part_id}_{ses_id}")
-
-    anat = {}
-    if isinstance(data_paths["anat"], str):
-        anat["T1"] = data_paths["anat"]
-    elif "T1w" in data_paths["anat"]:
-        anat["T1"] = data_paths["anat"]["T1w"]
-
-    if "T1" in anat:
-        anat_flow.inputs.inputnode.set(
-            subject=part_id,
-            anat=anat["T1"],
-            creds_path=data_paths["creds_path"],
-            dl_dir=cfg.pipeline_setup["working_directory"]["path"],
-            img_type="anat",
-        )
-        rpool.set_data("T1w", anat_flow, "outputspec.anat", {}, "", "anat_ingress")
-
-    if "T2w" in data_paths["anat"]:
-        anat_flow_T2 = create_anat_datasource(f"anat_T2w_gather_{part_id}_{ses_id}")
-        anat_flow_T2.inputs.inputnode.set(
-            subject=part_id,
-            anat=data_paths["anat"]["T2w"],
-            creds_path=data_paths["creds_path"],
-            dl_dir=cfg.pipeline_setup["working_directory"]["path"],
-            img_type="anat",
-        )
-        rpool.set_data("T2w", anat_flow_T2, "outputspec.anat", {}, "", "anat_ingress")
-
-    if cfg.surface_analysis["freesurfer"]["ingress_reconall"]:
-        rpool = ingress_freesurfer(
-            wf, rpool, cfg, data_paths, unique_id, part_id, ses_id
-        )
-
-    return rpool
-
-
-def ingress_freesurfer(wf, rpool, cfg, data_paths, unique_id, part_id, ses_id):
-    try:
-        fs_path = os.path.join(cfg.pipeline_setup["freesurfer_dir"], part_id)
-    except KeyError:
-        WFLOGGER.warning("No FreeSurfer data present.")
-        return rpool
-
-    # fs_path = os.path.join(cfg.pipeline_setup['freesurfer_dir'], part_id)
-    if not os.path.exists(fs_path):
-        if "sub" in part_id:
-            fs_path = os.path.join(
-                cfg.pipeline_setup["freesurfer_dir"], part_id.replace("sub-", "")
-            )
-        else:
-            fs_path = os.path.join(
-                cfg.pipeline_setup["freesurfer_dir"], ("sub-" + part_id)
-            )
-
-        # patch for flo-specific data
-        if not os.path.exists(fs_path):
-            subj_ses = part_id + "-" + ses_id
-            fs_path = os.path.join(cfg.pipeline_setup["freesurfer_dir"], subj_ses)
-            if not os.path.exists(fs_path):
-                WFLOGGER.info("No FreeSurfer data found for subject %s", part_id)
-                return rpool
-
-    # Check for double nested subj names
-    if os.path.exists(os.path.join(fs_path, os.path.basename(fs_path))):
-        fs_path = os.path.join(fs_path, part_id)
-
-    fs_ingress = create_general_datasource("gather_freesurfer_dir")
-    fs_ingress.inputs.inputnode.set(
-        unique_id=unique_id,
-        data=fs_path,
-        creds_path=data_paths["creds_path"],
-        dl_dir=cfg.pipeline_setup["working_directory"]["path"],
-    )
-    rpool.set_data(
-        "freesurfer-subject-dir",
-        fs_ingress,
-        "outputspec.data",
-        {},
-        "",
-        "freesurfer_config_ingress",
-    )
-
-    recon_outs = {
-        "pipeline-fs_raw-average": "mri/rawavg.mgz",
-        "pipeline-fs_subcortical-seg": "mri/aseg.mgz",
-        "pipeline-fs_brainmask": "mri/brainmask.mgz",
-        "pipeline-fs_wmparc": "mri/wmparc.mgz",
-        "pipeline-fs_T1": "mri/T1.mgz",
-        "pipeline-fs_hemi-L_desc-surface_curv": "surf/lh.curv",
-        "pipeline-fs_hemi-R_desc-surface_curv": "surf/rh.curv",
-        "pipeline-fs_hemi-L_desc-surfaceMesh_pial": "surf/lh.pial",
-        "pipeline-fs_hemi-R_desc-surfaceMesh_pial": "surf/rh.pial",
-        "pipeline-fs_hemi-L_desc-surfaceMesh_smoothwm": "surf/lh.smoothwm",
-        "pipeline-fs_hemi-R_desc-surfaceMesh_smoothwm": "surf/rh.smoothwm",
-        "pipeline-fs_hemi-L_desc-surfaceMesh_sphere": "surf/lh.sphere",
-        "pipeline-fs_hemi-R_desc-surfaceMesh_sphere": "surf/rh.sphere",
-        "pipeline-fs_hemi-L_desc-surfaceMap_sulc": "surf/lh.sulc",
-        "pipeline-fs_hemi-R_desc-surfaceMap_sulc": "surf/rh.sulc",
-        "pipeline-fs_hemi-L_desc-surfaceMap_thickness": "surf/lh.thickness",
-        "pipeline-fs_hemi-R_desc-surfaceMap_thickness": "surf/rh.thickness",
-        "pipeline-fs_hemi-L_desc-surfaceMap_volume": "surf/lh.volume",
-        "pipeline-fs_hemi-R_desc-surfaceMap_volume": "surf/rh.volume",
-        "pipeline-fs_hemi-L_desc-surfaceMesh_white": "surf/lh.white",
-        "pipeline-fs_hemi-R_desc-surfaceMesh_white": "surf/rh.white",
-        "pipeline-fs_xfm": "mri/transforms/talairach.lta",
-    }
-
-    for key, outfile in recon_outs.items():
-        fullpath = os.path.join(fs_path, outfile)
-        if os.path.exists(fullpath):
-            fs_ingress = create_general_datasource(f"gather_fs_{key}_dir")
-            fs_ingress.inputs.inputnode.set(
-                unique_id=unique_id,
-                data=fullpath,
-                creds_path=data_paths["creds_path"],
-                dl_dir=cfg.pipeline_setup["working_directory"]["path"],
-            )
-            rpool.set_data(
-                key, fs_ingress, "outputspec.data", {}, "", f"fs_{key}_ingress"
-            )
-        else:
-            warnings.warn(
-                str(LookupError(f"\n[!] Path does not exist for {fullpath}.\n"))
-            )
-
-    return rpool
-
-
-def ingress_raw_func_data(wf, rpool, cfg, data_paths, unique_id, part_id, ses_id):
-    func_paths_dct = data_paths["func"]
-
-    func_wf = create_func_datasource(
-        func_paths_dct, rpool, f"func_ingress_{part_id}_{ses_id}"
-    )
-    func_wf.inputs.inputnode.set(
-        subject=part_id,
-        creds_path=data_paths["creds_path"],
-        dl_dir=cfg.pipeline_setup["working_directory"]["path"],
-    )
-    func_wf.get_node("inputnode").iterables = ("scan", list(func_paths_dct.keys()))
-
-    rpool.set_data("subject", func_wf, "outputspec.subject", {}, "", "func_ingress")
-    rpool.set_data("bold", func_wf, "outputspec.rest", {}, "", "func_ingress")
-    rpool.set_data("scan", func_wf, "outputspec.scan", {}, "", "func_ingress")
-    rpool.set_data(
-        "scan-params", func_wf, "outputspec.scan_params", {}, "", "scan_params_ingress"
-    )
-
-    # TODO: CHECK FOR PARAMETERS
-
-    wf, rpool, diff, blip, fmap_rp_list = ingress_func_metadata(
-        wf, cfg, rpool, data_paths, part_id, data_paths["creds_path"], ses_id
-    )
-
-    # Memoize list of local functional scans
-    # TODO: handle S3 files
-    # Skip S3 files for now
-
-    local_func_scans = [
-        func_paths_dct[scan]["scan"]
-        for scan in func_paths_dct.keys()
-        if not func_paths_dct[scan]["scan"].startswith("s3://")
-    ]
-    if local_func_scans:
-        # pylint: disable=protected-access
-        wf._local_func_scans = local_func_scans
-        if cfg.pipeline_setup["Debugging"]["verbose"]:
-            verbose_logger = getLogger("CPAC.engine")
-            verbose_logger.debug("local_func_scans: %s", local_func_scans)
-    del local_func_scans
-
-    return (wf, rpool, diff, blip, fmap_rp_list)
-
-
-def ingress_output_dir(
-    wf, cfg, rpool, unique_id, data_paths, part_id, ses_id, creds_path=None
-):
-    dir_path = data_paths["derivatives_dir"]
-
-    WFLOGGER.info("\nPulling outputs from %s.\n", dir_path)
-
-    anat = os.path.join(dir_path, "anat")
-    func = os.path.join(dir_path, "func")
-
-    exts = [".nii", ".gz", ".mat", ".1D", ".txt", ".csv", ".rms", ".tsv"]
-
-    outdir_anat = []
-    outdir_func = []
-    func_paths = {}
-    func_dict = {}
-
-    for subdir in [anat, func]:
-        if os.path.isdir(subdir):
-            for filename in os.listdir(subdir):
-                for ext in exts:
-                    if ext in filename:
-                        if subdir == anat:
-                            outdir_anat.append(os.path.join(subdir, filename))
-                        else:
-                            outdir_func.append(os.path.join(subdir, filename))
-
-    # Add derivatives directory to rpool
-    ingress = create_general_datasource("gather_derivatives_dir")
-    ingress.inputs.inputnode.set(
-        unique_id=unique_id,
-        data=dir_path,
-        creds_path=creds_path,
-        dl_dir=cfg.pipeline_setup["working_directory"]["path"],
-    )
-    rpool.set_data(
-        "derivatives-dir", ingress, "outputspec.data", {}, "", "outdir_config_ingress"
-    )
-
-    for subdir in [outdir_anat, outdir_func]:
-        for filepath in subdir:
-            filename = str(filepath)
-            for ext in exts:
-                filename = filename.split("/")[-1].replace(ext, "")
-
-            data_label = filename.split(unique_id)[1].lstrip("_")
-
-            if len(filename) == len(data_label):
-                msg = (
-                    "\n\n[!] Possibly wrong participant or "
-                    "session in this directory?\n\n"
-                    f"Filepath: {filepath}\n\n"
-                )
-                raise Exception(msg)
-
-            bidstag = ""
-            for tag in data_label.split("_"):
-                for prefix in ["task-", "run-", "acq-", "rec"]:
-                    if tag.startswith(prefix):
-                        bidstag += f"{tag}_"
-                        data_label = data_label.replace(f"{tag}_", "")
-            data_label, json = strip_template(data_label, dir_path, filename)
-
-            rpool, json_info, pipe_idx, node_name, data_label = json_outdir_ingress(
-                rpool, filepath, exts, data_label, json
-            )
-
-            if (
-                "template" in data_label
-                and not json_info["Template"]
-                == cfg.pipeline_setup["outdir_ingress"]["Template"]
-            ):
-                continue
-            # Rename confounds to avoid confusion in nuisance regression
-            if data_label.endswith("desc-confounds_timeseries"):
-                data_label = "pipeline-ingress_desc-confounds_timeseries"
-
-            if len(bidstag) > 1:
-                # Remove tail symbol
-                bidstag = bidstag[:-1]
-                if bidstag.startswith("task-"):
-                    bidstag = bidstag.replace("task-", "")
-
-            # Rename bold mask for CPAC naming convention
-            # and to avoid collision with anat brain mask
-            if data_label.endswith("desc-brain_mask") and filepath in outdir_func:
-                data_label = data_label.replace("brain_mask", "bold_mask")
-
-            try:
-                pipe_x = rpool.get_pipe_number(pipe_idx)
-            except ValueError:
-                pipe_x = len(rpool.pipe_list)
-            if filepath in outdir_anat:
-                ingress = create_general_datasource(
-                    f"gather_anat_outdir_{data_label!s}_{pipe_x}"
-                )
-                ingress.inputs.inputnode.set(
-                    unique_id=unique_id,
-                    data=filepath,
-                    creds_path=creds_path,
-                    dl_dir=cfg.pipeline_setup["working_directory"]["path"],
-                )
-                rpool.set_data(
-                    data_label,
-                    ingress,
-                    "outputspec.data",
-                    json_info,
-                    pipe_idx,
-                    node_name,
-                    f"outdir_{data_label}_ingress",
-                    inject=True,
-                )
-            else:
-                if data_label.endswith("desc-preproc_bold"):
-                    func_key = data_label
-                    func_dict[bidstag] = {}
-                    func_dict[bidstag]["scan"] = str(filepath)
-                    func_dict[bidstag]["scan_parameters"] = json_info
-                    func_dict[bidstag]["pipe_idx"] = pipe_idx
-                if data_label.endswith("desc-brain_mask"):
-                    data_label = data_label.replace("brain_mask", "bold_mask")
-                try:
-                    func_paths[data_label].append(filepath)
-                except:
-                    func_paths[data_label] = []
-                    func_paths[data_label].append(filepath)
-
-    if func_dict:
-        wf, rpool = func_outdir_ingress(
-            wf,
-            cfg,
-            func_dict,
-            rpool,
-            unique_id,
-            creds_path,
-            part_id,
-            func_key,
-            func_paths,
-        )
-
-    if cfg.surface_analysis["freesurfer"]["ingress_reconall"]:
-        rpool = ingress_freesurfer(
-            wf, rpool, cfg, data_paths, unique_id, part_id, ses_id
-        )
-    return wf, rpool
-
-
-def json_outdir_ingress(rpool, filepath, exts, data_label, json):
-    desc_val = None
-    for tag in data_label.split("_"):
-        if "desc-" in tag:
-            desc_val = tag
-            break
-    jsonpath = str(filepath)
-    for ext in exts:
-        jsonpath = jsonpath.replace(ext, "")
-    jsonpath = f"{jsonpath}.json"
-
-    if not os.path.exists(jsonpath):
-        WFLOGGER.info(
-            "\n\n[!] No JSON found for file %s.\nCreating %s..\n\n", filepath, jsonpath
-        )
-        json_info = {
-            "Description": "This data was generated elsewhere and "
-            "supplied by the user into this C-PAC run's "
-            "output directory. This JSON file was "
-            "automatically generated by C-PAC because a "
-            "JSON file was not supplied with the data."
-        }
-        json_info = {**json_info, **json}
-        write_output_json(json_info, jsonpath)
-    else:
-        json_info = read_json(jsonpath)
-        json_info = {**json_info, **json}
-    if "CpacProvenance" in json_info:
-        if desc_val:
-            # it's a C-PAC output, let's check for pipe_idx/strat integer
-            # suffixes in the desc- entries.
-            only_desc = str(desc_val)
-
-            if only_desc[-1].isdigit():
-                for idx in range(0, 3):
-                    # let's stop at 3, please don't run >999 strategies okay?
-                    if only_desc[-1].isdigit():
-                        only_desc = only_desc[:-1]
-
-                if only_desc[-1] == "-":
-                    only_desc = only_desc.rstrip("-")
-                else:
-                    msg = (
-                        "\n[!] Something went wrong with either "
-                        "reading in the output directory or when "
-                        "it was written out previously.\n\nGive "
-                        "this to your friendly local C-PAC "
-                        f"developer:\n\n{data_label!s}\n"
-                    )
-                    raise IOError(msg)
-
-            # remove the integer at the end of the desc-* variant, we will
-            # get the unique pipe_idx from the CpacProvenance below
-            data_label = data_label.replace(desc_val, only_desc)
-
-        # preserve cpac provenance/pipe_idx
-        pipe_idx = rpool.generate_prov_string(json_info["CpacProvenance"])
-        node_name = ""
-
-    else:
-        json_info["CpacProvenance"] = [f"{data_label}:Non-C-PAC Origin: {filepath}"]
-        if "Description" not in json_info:
-            json_info["Description"] = (
-                "This data was generated elsewhere and "
-                "supplied by the user into this C-PAC run's "
-                "output directory. This JSON file was "
-                "automatically generated by C-PAC because a "
-                "JSON file was not supplied with the data."
-            )
-        pipe_idx = rpool.generate_prov_string(json_info["CpacProvenance"])
-        node_name = f"{data_label}_ingress"
-
-    return rpool, json_info, pipe_idx, node_name, data_label
-
-
-def func_outdir_ingress(
-    wf, cfg, func_dict, rpool, unique_id, creds_path, part_id, key, func_paths
-):
-    pipe_x = len(rpool.pipe_list)
-    ingress = create_func_datasource(
-        func_dict, rpool, f"gather_func_outdir_{key}_{pipe_x}"
-    )
-    ingress.inputs.inputnode.set(
-        subject=unique_id,
-        creds_path=creds_path,
-        dl_dir=cfg.pipeline_setup["working_directory"]["path"],
-    )
-    rpool.set_data("subject", ingress, "outputspec.subject", {}, "", "func_ingress")
-    ingress.get_node("inputnode").iterables = ("scan", list(func_dict.keys()))
-    rpool.set_data(key, ingress, "outputspec.rest", {}, "", "func_ingress")
-
-    rpool.set_data("scan", ingress, "outputspec.scan", {}, "", "func_ingress")
-    rpool.set_data(
-        "scan-params", ingress, "outputspec.scan_params", {}, "", "scan_params_ingress"
-    )
-    wf, rpool, diff, blip, fmap_rp_list = ingress_func_metadata(
-        wf, cfg, rpool, func_dict, part_id, creds_path, key
-    )
-
-    # Have to do it this weird way to save the parsed BIDS tag & filepath
-    mask_paths_key = (
-        "desc-bold_mask"
-        if "desc-bold_mask" in func_paths
-        else "space-template_desc-bold_mask"
-    )
-    ts_paths_key = "pipeline-ingress_desc-confounds_timeseries"
-
-    # Connect func data with approproate scan name
-    iterables = pe.Node(
-        Function(
-            input_names=["scan", "mask_paths", "ts_paths"],
-            output_names=["out_scan", "mask", "confounds"],
-            function=set_iterables,
-        ),
-        name=f"set_iterables_{pipe_x}",
-    )
-    iterables.inputs.mask_paths = func_paths[mask_paths_key]
-    iterables.inputs.ts_paths = func_paths[ts_paths_key]
-    wf.connect(ingress, "outputspec.scan", iterables, "scan")
-
-    for key in func_paths:
-        if key in (mask_paths_key, ts_paths_key):
-            ingress_func = create_general_datasource(f"ingress_func_data_{key}")
-            ingress_func.inputs.inputnode.set(
-                unique_id=unique_id,
-                creds_path=creds_path,
-                dl_dir=cfg.pipeline_setup["working_directory"]["path"],
-            )
-            wf.connect(iterables, "out_scan", ingress_func, "inputnode.scan")
-            if key == mask_paths_key:
-                wf.connect(iterables, "mask", ingress_func, "inputnode.data")
-                rpool.set_data(
-                    key, ingress_func, "inputnode.data", {}, "", f"outdir_{key}_ingress"
-                )
-            elif key == ts_paths_key:
-                wf.connect(iterables, "confounds", ingress_func, "inputnode.data")
-                rpool.set_data(
-                    key, ingress_func, "inputnode.data", {}, "", f"outdir_{key}_ingress"
-                )
-
-    return wf, rpool
-
-
-def set_iterables(scan, mask_paths=None, ts_paths=None):
-    # match scan with filepath to get filepath
-    mask_path = [path for path in mask_paths if scan in path]
-    ts_path = [path for path in ts_paths if scan in path]
-
-    return (scan, mask_path[0], ts_path[0])
-
-
-def strip_template(data_label, dir_path, filename):
-    json = {}
-    # rename to template
-    for prefix in ["space-", "from-", "to-"]:
-        for bidstag in data_label.split("_"):
-            if bidstag.startswith(prefix):
-                template_key, template_val = bidstag.split("-")
-                template_name, _template_desc = lookup_identifier(template_val)
-                if template_name:
-                    json["Template"] = template_val
-                    data_label = data_label.replace(template_val, "template")
-            elif bidstag.startswith("res-"):
-                res_key, res_val = bidstag.split("-")
-                json["Resolution"] = res_val
-                data_label = data_label.replace(bidstag, "")
-    if data_label.find("__"):
-        data_label = data_label.replace("__", "_")
-    return data_label, json
-
-
-def ingress_pipeconfig_paths(cfg, rpool, unique_id, creds_path=None):
-    # ingress config file paths
-    # TODO: may want to change the resource keys for each to include one level up in the YAML as well
-
-    import pandas as pd
-    import pkg_resources as p
-
-    template_csv = p.resource_filename("CPAC", "resources/cpac_templates.csv")
-    template_df = pd.read_csv(template_csv, keep_default_na=False)
-
-    for row in template_df.itertuples():
-        key = row.Key
-        val = row.Pipeline_Config_Entry
-        val = cfg.get_nested(cfg, [x.lstrip() for x in val.split(",")])
-        resolution = row.Intended_Resolution_Config_Entry
-        desc = row.Description
-
-        if not val:
-            continue
-
-        if resolution:
-            res_keys = [x.lstrip() for x in resolution.split(",")]
-            tag = res_keys[-1]
-        json_info = {}
-
-        if "$FSLDIR" in val:
-            val = val.replace("$FSLDIR", cfg.pipeline_setup["system_config"]["FSLDIR"])
-        if "$priors_path" in val:
-            priors_path = (
-                cfg.segmentation["tissue_segmentation"]["FSL-FAST"]["use_priors"][
-                    "priors_path"
-                ]
-                or ""
-            )
-            if "$FSLDIR" in priors_path:
-                priors_path = priors_path.replace(
-                    "$FSLDIR", cfg.pipeline_setup["system_config"]["FSLDIR"]
-                )
-            val = val.replace("$priors_path", priors_path)
-        if "${resolution_for_anat}" in val:
-            val = val.replace(
-                "${resolution_for_anat}",
-                cfg.registration_workflows["anatomical_registration"][
-                    "resolution_for_anat"
-                ],
-            )
-        if "${func_resolution}" in val:
-            val = val.replace(
-                "${func_resolution}",
-                cfg.registration_workflows["functional_registration"][
-                    "func_registration_to_template"
-                ]["output_resolution"][tag],
-            )
-
-        if desc:
-            template_name, _template_desc = lookup_identifier(val)
-            if template_name:
-                desc = f"{template_name} - {desc}"
-            json_info["Description"] = f"{desc} - {val}"
-        if resolution:
-            resolution = cfg.get_nested(cfg, res_keys)
-            json_info["Resolution"] = resolution
-
-            resampled_template = pe.Node(
-                Function(
-                    input_names=["resolution", "template", "template_name", "tag"],
-                    output_names=["resampled_template"],
-                    function=resolve_resolution,
-                    as_module=True,
-                ),
-                name="resampled_" + key,
-            )
-
-            resampled_template.inputs.resolution = resolution
-            resampled_template.inputs.template = val
-            resampled_template.inputs.template_name = key
-            resampled_template.inputs.tag = tag
-
-            # the set_data below is set up a little differently, because we are
-            # injecting and also over-writing already-existing entries
-            # other alternative would have been to ingress into the
-            # resampled_template node from the already existing entries, but we
-            # didn't do that here
-            rpool.set_data(
-                key,
-                resampled_template,
-                "resampled_template",
-                json_info,
-                "",
-                "template_resample",
-            )  # pipe_idx (after the blank json {}) should be the previous strat that you want deleted! because you're not connecting this the regular way, you have to do it manually
-
-        elif val:
-            config_ingress = create_general_datasource(f"gather_{key}")
-            config_ingress.inputs.inputnode.set(
-                unique_id=unique_id,
-                data=val,
-                creds_path=creds_path,
-                dl_dir=cfg.pipeline_setup["working_directory"]["path"],
-            )
-            rpool.set_data(
-                key,
-                config_ingress,
-                "outputspec.data",
-                json_info,
-                "",
-                f"{key}_config_ingress",
-            )
-    # templates, resampling from config
-    """
-    template_keys = [
-        ("anat", ["network_centrality", "template_specification_file"]),
-        ("anat", ["nuisance_corrections", "2-nuisance_regression",
-                  "lateral_ventricles_mask"]),
-        ("anat",
-         ["segmentation", "tissue_segmentation", "FSL-FAST", "use_priors",
-          "CSF_path"]),
-        ("anat",
-         ["segmentation", "tissue_segmentation", "FSL-FAST", "use_priors",
-          "GM_path"]),
-        ("anat",
-         ["segmentation", "tissue_segmentation", "FSL-FAST", "use_priors",
-          "WM_path"]),
-        ("anat",
-         ["segmentation", "tissue_segmentation", "Template_Based", "CSF"]),
-        ("anat",
-         ["segmentation", "tissue_segmentation", "Template_Based", "GRAY"]),
-        ("anat",
-         ["segmentation", "tissue_segmentation", "Template_Based", "WHITE"]),
-        ("anat", ["anatomical_preproc", "acpc_alignment", "T1w_ACPC_template"]),
-        ("anat", ["anatomical_preproc", "acpc_alignment", "T1w_brain_ACPC_template"]),
-        ("anat", ["anatomical_preproc", "acpc_alignment", "T2w_ACPC_template"]),
-        ("anat", ["anatomical_preproc", "acpc_alignment", "T2w_brain_ACPC_template"])]
-
-    def get_nested_attr(c, template_key):
-        attr = getattr(c, template_key[0])
-        keys = template_key[1:]
-
-        def _get_nested(attr, keys):
-            if len(keys) > 1:
-                return (_get_nested(attr[keys[0]], keys[1:]))
-            elif len(keys):
-                return (attr[keys[0]])
-            else:
-                return (attr)
-
-        return (_get_nested(attr, keys))
-
-    def set_nested_attr(c, template_key, value):
-        attr = getattr(c, template_key[0])
-        keys = template_key[1:]
-
-        def _set_nested(attr, keys):
-            if len(keys) > 1:
-                return (_set_nested(attr[keys[0]], keys[1:]))
-            elif len(keys):
-                attr[keys[0]] = value
-            else:
-                return (attr)
-
-        return (_set_nested(attr, keys))
-
-    for key_type, key in template_keys:
-        attr = cfg.get_nested(cfg, key)
-        if isinstance(attr, str) or attr == None:
-            node = create_check_for_s3_node(
-                key[-1],
-                attr, key_type,
-                data_paths['creds_path'],
-                cfg.pipeline_setup['working_directory']['path'],
-                map_node=False
-            )
-            cfg.set_nested(cfg, key, node)
-
-    template_keys_in_list = [
-        ("anat",
-         ["segmentation", "tissue_segmentation", "ANTs_Prior_Based",
-          "template_brain_list"]),
-        ("anat",
-         ["segmentation", "tissue_segmentation", "ANTs_Prior_Based",
-          "template_segmentation_list"]),
-    ]
-
-    for key_type, key in template_keys_in_list:
-        node = create_check_for_s3_node(
-            key[-1],
-            cfg.get_nested(cfg, key), key_type,
-            data_paths['creds_path'],
-            cfg.pipeline_setup['working_directory']['path'],
-            map_node=True
-        )
-        cfg.set_nested(cfg, key, node)
-    """
-
-    return rpool
-
-
 def run_node_blocks(blocks, data_paths, cfg=None):
-    import os
-
-    from CPAC.pipeline import nipype_pipeline_engine as pe
     from CPAC.pipeline.engine import NodeBlock
     from CPAC.pipeline.engine.resource import initiate_rpool
 
@@ -1251,9 +538,9 @@ def run_node_blocks(blocks, data_paths, cfg=None):
         }
 
     # TODO: WE HAVE TO PARSE OVER UNIQUE ID'S!!!
-    _, rpool = initiate_rpool(cfg, data_paths)
 
     wf = pe.Workflow(name="node_blocks")
+    rpool = initiate_rpool(wf, cfg, data_paths)
     wf.base_dir = cfg.pipeline_setup["working_directory"]["path"]
     wf.config["execution"] = {
         "hash_method": "timestamp",
