+++ CPAC/pipeline/engine/resource.py
@@ -20,10 +20,12 @@ import ast
 import copy
 from itertools import chain
 import os
+from pathlib import Path
 import re
 from typing import Optional
 import warnings
 
+from nipype.interfaces import utility as util
 from nipype.interfaces.utility import Rename
 
 from CPAC.image_utils.spatial_smoothing import spatial_smoothing
@@ -33,16 +35,23 @@ from CPAC.image_utils.statistical_transforms import (
 )
 from CPAC.pipeline import nipype_pipeline_engine as pe
 from CPAC.pipeline.check_outputs import ExpectedOutputs
-from CPAC.pipeline.engine.engine import (
-    ingress_output_dir,
-    ingress_pipeconfig_paths,
-    ingress_raw_anat_data,
-    ingress_raw_func_data,
-)
 from CPAC.pipeline.utils import MOVEMENT_FILTER_KEYS, name_fork, source_set
 from CPAC.registration.registration import transform_derivative
+from CPAC.resources.templates.lookup_table import lookup_identifier
 from CPAC.utils.bids_utils import res_in_filename
 from CPAC.utils.configuration import Configuration
+from CPAC.utils.datasource import (
+    calc_delta_te_and_asym_ratio,
+    check_for_s3,
+    check_func_scan,
+    create_anat_datasource,
+    create_fmap_datasource,
+    create_general_datasource,
+    gather_echo_times,
+    get_fmap_phasediff_metadata,
+    get_rest,
+    resolve_resolution,
+)
 from CPAC.utils.interfaces.datasink import DataSink
 from CPAC.utils.interfaces.function import Function
 from CPAC.utils.monitoring import (
@@ -54,8 +63,142 @@ from CPAC.utils.utils import (
     check_prov_for_regtool,
     create_id_string,
     get_last_prov_entry,
+    get_scan_params,
+    read_json,
     write_output_json,
 )
+
+EXTS = [".nii", ".gz", ".mat", ".1D", ".txt", ".csv", ".rms", ".tsv"]
+
+
+def generate_prov_string(prov: list[str]) -> tuple[str, str]:
+    """Generate a string from a SINGLE RESOURCE'S dictionary of MULTIPLE PRECEDING RESOURCES (or single, if just one).
+
+    NOTE: this DOES NOT merge multiple resources!!! (i.e. for merging-strat pipe_idx generation).
+    """
+    if not isinstance(prov, list):
+        msg = (
+            "\n[!] Developer info: the CpacProvenance "
+            f"entry for {prov} has to be a list.\n"
+        )
+        raise TypeError(msg)
+    last_entry = get_last_prov_entry(prov)
+    resource = last_entry.split(":")[0]
+    return (resource, str(prov))
+
+
+def json_outdir_ingress(
+    filepath: Path | str, data_label: str, json: dict
+) -> tuple[dict, tuple[str, str], str, str]:
+    """Ingress sidecars from a BIDS derivatives directory."""
+    desc_val = None
+    for tag in data_label.split("_"):
+        if "desc-" in tag:
+            desc_val = tag
+            break
+    jsonpath = str(filepath)
+    for ext in EXTS:
+        jsonpath = jsonpath.replace(ext, "")
+    jsonpath = f"{jsonpath}.json"
+
+    if not os.path.exists(jsonpath):
+        WFLOGGER.info(
+            "\n\n[!] No JSON found for file %s.\nCreating %s..\n\n",
+            filepath,
+            jsonpath,
+        )
+        json_info = {
+            "Description": "This data was generated elsewhere and "
+            "supplied by the user into this C-PAC run's "
+            "output directory. This JSON file was "
+            "automatically generated by C-PAC because a "
+            "JSON file was not supplied with the data."
+        }
+        json_info = {**json_info, **json}
+        write_output_json(json_info, jsonpath)
+    else:
+        json_info = read_json(jsonpath)
+        json_info = {**json_info, **json}
+    if "CpacProvenance" in json_info:
+        if desc_val:
+            # it's a C-PAC output, let's check for pipe_idx/strat integer
+            # suffixes in the desc- entries.
+            only_desc = str(desc_val)
+
+            if only_desc[-1].isdigit():
+                for _strat_idx in range(0, 3):
+                    # let's stop at 3, please don't run >999 strategies okay?
+                    if only_desc[-1].isdigit():
+                        only_desc = only_desc[:-1]
+
+                if only_desc[-1] == "-":
+                    only_desc = only_desc.rstrip("-")
+                else:
+                    msg = (
+                        "\n[!] Something went wrong with either "
+                        "reading in the output directory or when "
+                        "it was written out previously.\n\nGive "
+                        "this to your friendly local C-PAC "
+                        f"developer:\n\n{data_label!s}\n"
+                    )
+                    raise IOError(msg)
+
+            # remove the integer at the end of the desc-* variant, we will
+            # get the unique pipe_idx from the CpacProvenance below
+            data_label = data_label.replace(desc_val, only_desc)
+
+        # preserve cpac provenance/pipe_idx
+        pipe_idx = generate_prov_string(json_info["CpacProvenance"])
+        node_name = ""
+
+    else:
+        json_info["CpacProvenance"] = [f"{data_label}:Non-C-PAC Origin: {filepath}"]
+        if "Description" not in json_info:
+            json_info["Description"] = (
+                "This data was generated elsewhere and "
+                "supplied by the user into this C-PAC run's "
+                "output directory. This JSON file was "
+                "automatically generated by C-PAC because a "
+                "JSON file was not supplied with the data."
+            )
+        pipe_idx = generate_prov_string(json_info["CpacProvenance"])
+        node_name = f"{data_label}_ingress"
+
+    return json_info, pipe_idx, node_name, data_label
+
+
+@Function.sig_imports(["from typing import Optional"])
+def set_iterables(
+    scan: str,
+    mask_paths: Optional[list[str]] = None,
+    ts_paths: Optional[list[str]] = None,
+) -> tuple[str, str, str]:
+    """Match scan with filepath to get filepath."""
+    mask_path = [path for path in mask_paths if scan in path]
+    ts_path = [path for path in ts_paths if scan in path]
+
+    return (scan, mask_path[0], ts_path[0])
+
+
+def strip_template(data_label: str) -> tuple[str, dict[str, str]]:
+    """Strip a template name from a data label to use as a Resource key."""
+    json = {}
+    # rename to template
+    for prefix in ["space-", "from-", "to-"]:
+        for bidstag in data_label.split("_"):
+            if bidstag.startswith(prefix):
+                _template_key, template_val = bidstag.split("-")
+                template_name, _template_desc = lookup_identifier(template_val)
+                if template_name:
+                    json["Template"] = template_val
+                    data_label = data_label.replace(template_val, "template")
+            elif bidstag.startswith("res-"):
+                _res_key, res_val = bidstag.split("-")
+                json["Resolution"] = res_val
+                data_label = data_label.replace(bidstag, "")
+    if data_label.find("__"):
+        data_label = data_label.replace("__", "_")
+    return data_label, json
 
 
 class NodeData:
@@ -108,7 +251,28 @@ class Resource:
 class ResourcePool:
     """All Resources."""
 
-    def __init__(self, rpool=None, name=None, cfg=None, pipe_list=None):
+    def __init__(
+        self,
+        rpool: Optional[dict] = None,
+        name: str = "",
+        cfg: Optional[Configuration] = None,
+        pipe_list: Optional[list] = None,
+        *,
+        creds_path: Optional[str] = None,
+        data_paths: Optional[dict] = None,
+        part_id: Optional[str] = None,
+        ses_id: Optional[str] = None,
+        unique_id: Optional[str] = None,
+        wf: Optional[pe.Workflow] = None,
+        **kwargs,
+    ):
+        """Initialize a ResourcePool."""
+        self.creds_path = creds_path
+        self.data_paths = data_paths
+        self.part_id = part_id
+        self.ses_id = ses_id
+        self.unique_id = unique_id
+        self._init_wf = wf
         if not rpool:
             self.rpool = {}
         else:
@@ -171,23 +335,22 @@ class ResourcePool:
         ]
 
     def __repr__(self) -> str:
+        """Return reproducible ResourcePool string."""
         params = [
             f"{param}={getattr(self, param)}"
             for param in ["rpool", "name", "cfg", "pipe_list"]
-            if getattr(self, param, None) is not None
+            if getattr(self, param, None)
         ]
         return f'ResourcePool({", ".join(params)})'
 
     def __str__(self) -> str:
+        """Return string representation of ResourcePool."""
         if self.name:
             return f"ResourcePool({self.name}): {list(self.rpool)}"
         return f"ResourcePool: {list(self.rpool)}"
 
-    def append_name(self, name):
-        self.name.append(name)
-
     def back_propogate_template_name(
-        self, wf, resource_idx: str, json_info: dict, id_string: "pe.Node"
+        self, resource_idx: str, json_info: dict, id_string: "pe.Node"
     ) -> None:
         """Find and apply the template name from a resource's provenance.
 
@@ -206,7 +369,7 @@ class ResourcePool:
         if "template" in resource_idx and self.check_rpool("derivatives-dir"):
             if self.check_rpool("template"):
                 node, out = self.get_data("template")
-                wf.connect(node, out, id_string, "template_desc")
+                self._init_wf.connect(node, out, id_string, "template_desc")
         elif "Template" in json_info:
             id_string.inputs.template_desc = json_info["Template"]
         elif (
@@ -241,6 +404,125 @@ class ResourcePool:
             if name in self.rpool:
                 return True
         return False
+
+    def create_func_datasource(
+        self, rest_dict: dict, wf_name="func_datasource"
+    ) -> pe.Workflow:
+        """Create a workflow to gather timeseries data.
+
+        Return the functional timeseries-related file paths for each series/scan from the
+        dictionary of functional files described in the data configuration (sublist) YAML
+        file.
+
+        Scan input (from inputnode) is an iterable.
+        """
+        wf = pe.Workflow(name=wf_name)
+
+        inputnode = pe.Node(
+            util.IdentityInterface(
+                fields=["subject", "scan", "creds_path", "dl_dir"],
+                mandatory_inputs=True,
+            ),
+            name="inputnode",
+        )
+
+        outputnode = pe.Node(
+            util.IdentityInterface(
+                fields=[
+                    "subject",
+                    "rest",
+                    "scan",
+                    "scan_params",
+                    "phase_diff",
+                    "magnitude",
+                ]
+            ),
+            name="outputspec",
+        )
+
+        # have this here for now because of the big change in the data
+        # configuration format
+        # (Not necessary with ingress - format does not comply)
+        if not self.check_rpool("derivatives-dir"):
+            check_scan = pe.Node(
+                Function(
+                    input_names=["func_scan_dct", "scan"],
+                    output_names=[],
+                    function=check_func_scan,
+                    as_module=True,
+                ),
+                name="check_func_scan",
+            )
+
+            check_scan.inputs.func_scan_dct = rest_dict
+            wf.connect(inputnode, "scan", check_scan, "scan")
+
+        # get the functional scan itself
+        selectrest = pe.Node(
+            Function(
+                input_names=["scan", "rest_dict", "resource"],
+                output_names=["file_path"],
+                function=get_rest,
+                as_module=True,
+            ),
+            name="selectrest",
+        )
+        selectrest.inputs.rest_dict = rest_dict
+        selectrest.inputs.resource = "scan"
+        wf.connect(inputnode, "scan", selectrest, "scan")
+
+        # check to see if it's on an Amazon AWS S3 bucket, and download it, if it
+        # is - otherwise, just return the local file path
+        check_s3_node = pe.Node(
+            Function(
+                input_names=["file_path", "creds_path", "dl_dir", "img_type"],
+                output_names=["local_path"],
+                function=check_for_s3,
+                as_module=True,
+            ),
+            name="check_for_s3",
+        )
+
+        wf.connect(selectrest, "file_path", check_s3_node, "file_path")
+        wf.connect(inputnode, "creds_path", check_s3_node, "creds_path")
+        wf.connect(inputnode, "dl_dir", check_s3_node, "dl_dir")
+        check_s3_node.inputs.img_type = "func"
+
+        wf.connect(inputnode, "subject", outputnode, "subject")
+        wf.connect(check_s3_node, "local_path", outputnode, "rest")
+        wf.connect(inputnode, "scan", outputnode, "scan")
+
+        # scan parameters CSV
+        select_scan_params = pe.Node(
+            Function(
+                input_names=["scan", "rest_dict", "resource"],
+                output_names=["file_path"],
+                function=get_rest,
+                as_module=True,
+            ),
+            name="select_scan_params",
+        )
+        select_scan_params.inputs.rest_dict = rest_dict
+        select_scan_params.inputs.resource = "scan_parameters"
+        wf.connect(inputnode, "scan", select_scan_params, "scan")
+
+        # if the scan parameters file is on AWS S3, download it
+        s3_scan_params = pe.Node(
+            Function(
+                input_names=["file_path", "creds_path", "dl_dir", "img_type"],
+                output_names=["local_path"],
+                function=check_for_s3,
+                as_module=True,
+            ),
+            name="s3_scan_params",
+        )
+
+        wf.connect(select_scan_params, "file_path", s3_scan_params, "file_path")
+        wf.connect(inputnode, "creds_path", s3_scan_params, "creds_path")
+        wf.connect(inputnode, "dl_dir", s3_scan_params, "dl_dir")
+        wf.connect(s3_scan_params, "local_path", outputnode, "scan_params")
+
+        return wf
 
     def get_pipe_number(self, pipe_idx):
         return self.pipe_list.index(pipe_idx)
@@ -382,7 +664,7 @@ class ResourcePool:
         if not inject:
             new_prov_list.append(f"{resource}:{node_name}")
         try:
-            res, new_pipe_idx = self.generate_prov_string(new_prov_list)
+            _res, new_pipe_idx = generate_prov_string(new_prov_list)
         except IndexError:
             msg = (
                 f"\n\nThe set_data() call for {resource} has no "
@@ -404,7 +686,7 @@ class ResourcePool:
             search = False
             if self.get_resource_from_prov(current_prov_list) == resource:
                 # CHANGING PIPE_IDX, BE CAREFUL DOWNSTREAM IN THIS FUNCTION
-                pipe_idx = self.generate_prov_string(current_prov_list)[1]
+                pipe_idx = generate_prov_string(current_prov_list)[1]
                 if pipe_idx not in self.rpool[resource].keys():
                     search = True
             else:
@@ -414,7 +696,7 @@ class ResourcePool:
                     if self.get_resource_from_prov(idx) == resource:
                         if isinstance(idx, list):
                             # CHANGING PIPE_IDX, BE CAREFUL DOWNSTREAM IN THIS FUNCTION
-                            pipe_idx = self.generate_prov_string(idx)[1]
+                            pipe_idx = generate_prov_string(idx)[1]
                         elif isinstance(idx, str):
                             pipe_idx = idx
                         break
@@ -536,21 +818,6 @@ class ResourcePool:
                     continue
         json_data = self.get_json(resource, strat)
         return json_data["CpacProvenance"]
-
-    @staticmethod
-    def generate_prov_string(prov):
-        # this will generate a string from a SINGLE RESOURCE'S dictionary of
-        # MULTIPLE PRECEDING RESOURCES (or single, if just one)
-        # NOTE: this DOES NOT merge multiple resources!!! (i.e. for merging-strat pipe_idx generation)
-        if not isinstance(prov, list):
-            msg = (
-                "\n[!] Developer info: the CpacProvenance "
-                f"entry for {prov} has to be a list.\n"
-            )
-            raise TypeError(msg)
-        last_entry = get_last_prov_entry(prov)
-        resource = last_entry.split(":")[0]
-        return (resource, str(prov))
 
     @staticmethod
     def generate_prov_list(prov_str):
@@ -704,7 +971,7 @@ class ResourcePool:
                 json_dct = {}
                 for strat in strat_list:
                     # strat is a prov list for a single resource/input
-                    strat_resource, strat_idx = self.generate_prov_string(strat)
+                    strat_resource, strat_idx = generate_prov_string(strat)
                     strat_json = self.get_json(strat_resource, strat=strat_idx)
                     json_dct[strat_resource] = strat_json
 
@@ -797,7 +1064,7 @@ class ResourcePool:
 
                 # now just invert resource:strat to strat:resource for each resource:strat
                 for cpac_prov in strat_list:
-                    resource, strat = self.generate_prov_string(cpac_prov)
+                    resource, strat = generate_prov_string(cpac_prov)
                     resource_strat_dct = self.rpool[resource][strat]
                     # remember, `resource_strat_dct` is the dct of 'data' and 'json'.
                     new_strats[pipe_idx].rpool[resource] = resource_strat_dct
@@ -828,7 +1095,7 @@ class ResourcePool:
             for resource_strat_list in total_pool:
                 # total_pool will have only one list of strats, for the one input
                 for cpac_prov in resource_strat_list:  # <------- cpac_prov here doesn't need to be modified, because it's not merging with other inputs
-                    resource, pipe_idx = self.generate_prov_string(cpac_prov)
+                    resource, pipe_idx = generate_prov_string(cpac_prov)
                     resource_strat_dct = self.rpool[resource][pipe_idx]
                     # remember, `resource_strat_dct` is the dct of 'data' and 'json'.
                     new_strats[pipe_idx] = ResourcePool(
@@ -859,7 +1126,7 @@ class ResourcePool:
                 if isinstance(entry, list):
                     if entry[-1].split(":")[0] == xfm_label:
                         xfm_prov = entry
-                        xfm_idx = self.generate_prov_string(xfm_prov)[1]
+                        xfm_idx = generate_prov_string(xfm_prov)[1]
                         break
 
             # but if the resource doesn't have the bold-to-template transform
@@ -902,7 +1169,7 @@ class ResourcePool:
                 )
                 new_prov = json_info["CpacProvenance"] + xfm_prov
                 json_info["CpacProvenance"] = new_prov
-                new_pipe_idx = self.generate_prov_string(new_prov)
+                new_pipe_idx = generate_prov_string(new_prov)
                 self.set_data(
                     label,
                     xfm,
@@ -986,7 +1253,7 @@ class ResourcePool:
             if isinstance(entry, list):
                 if entry[-1].split(":")[0] == mask:
                     mask_prov = entry
-                    mask_idx = self.generate_prov_string(mask_prov)[1]
+                    mask_idx = generate_prov_string(mask_prov)[1]
                     break
 
         if self.smoothing_bool:
@@ -1110,7 +1377,7 @@ class ResourcePool:
     def gather_pipes(self, wf, cfg, all=False, add_incl=None, add_excl=None):
         excl = []
         substring_excl = []
-        outputs_logger = getLogger(f'{cfg["subject_id"]}_expectedOutputs')
+        outputs_logger = getLogger(f"{self.part_id}_expectedOutputs")
         expected_outputs = ExpectedOutputs()
 
         if add_excl:
@@ -1333,9 +1600,7 @@ class ResourcePool:
                     node, out = self.rpool["scan"]["['scan:func_ingress']"]["data"]
                     wf.connect(node, out, id_string, "scan_id")
 
-                self.back_propogate_template_name(
-                    wf, resource_idx, json_info, id_string
-                )
+                self.back_propogate_template_name(resource_idx, json_info, id_string)
                 # grab the FWHM if smoothed
                 for tag in resource.split("_"):
                     if "desc-" in tag and "-sm" in tag:
@@ -1453,12 +1718,846 @@ class ResourcePool:
         """
         return NodeData(self, resource, **kwargs)
 
+    def ingress_freesurfer(self) -> None:
+        """Ingress FreeSurfer data."""
+        try:
+            fs_path = os.path.join(
+                self.cfg.pipeline_setup["freesurfer_dir"], self.part_id
+            )
+        except KeyError:
+            WFLOGGER.warning("No FreeSurfer data present.")
+            return
 
-def initiate_rpool(wf, cfg, data_paths=None, part_id=None):
+        # fs_path = os.path.join(cfg.pipeline_setup['freesurfer_dir'], part_id)
+        if not os.path.exists(fs_path):
+            if "sub" in self.part_id:
+                fs_path = os.path.join(
+                    self.cfg.pipeline_setup["freesurfer_dir"],
+                    self.part_id.replace("sub-", ""),
+                )
+            else:
+                fs_path = os.path.join(
+                    self.cfg.pipeline_setup["freesurfer_dir"], ("sub-" + self.part_id)
+                )
+
+            # patch for flo-specific data
+            if not os.path.exists(fs_path):
+                subj_ses = f"{self.part_id}-{self.ses_id}"
+                fs_path = os.path.join(
+                    self.cfg.pipeline_setup["freesurfer_dir"], subj_ses
+                )
+                if not os.path.exists(fs_path):
+                    WFLOGGER.info(
+                        "No FreeSurfer data found for subject %s", self.part_id
+                    )
+                    return
+
+        # Check for double nested subj names
+        if os.path.exists(os.path.join(fs_path, os.path.basename(fs_path))):
+            fs_path = os.path.join(fs_path, self.part_id)
+
+        fs_ingress = create_general_datasource("gather_freesurfer_dir")
+        fs_ingress.inputs.inputnode.set(
+            unique_id=self.unique_id,
+            data=fs_path,
+            creds_path=self.creds_path,
+            dl_dir=self.cfg.pipeline_setup["working_directory"]["path"],
+        )
+        self.set_data(
+            "freesurfer-subject-dir",
+            fs_ingress,
+            "outputspec.data",
+            {},
+            "",
+            "freesurfer_config_ingress",
+        )
+
+        recon_outs = {
+            "pipeline-fs_raw-average": "mri/rawavg.mgz",
+            "pipeline-fs_subcortical-seg": "mri/aseg.mgz",
+            "pipeline-fs_brainmask": "mri/brainmask.mgz",
+            "pipeline-fs_wmparc": "mri/wmparc.mgz",
+            "pipeline-fs_T1": "mri/T1.mgz",
+            "pipeline-fs_hemi-L_desc-surface_curv": "surf/lh.curv",
+            "pipeline-fs_hemi-R_desc-surface_curv": "surf/rh.curv",
+            "pipeline-fs_hemi-L_desc-surfaceMesh_pial": "surf/lh.pial",
+            "pipeline-fs_hemi-R_desc-surfaceMesh_pial": "surf/rh.pial",
+            "pipeline-fs_hemi-L_desc-surfaceMesh_smoothwm": "surf/lh.smoothwm",
+            "pipeline-fs_hemi-R_desc-surfaceMesh_smoothwm": "surf/rh.smoothwm",
+            "pipeline-fs_hemi-L_desc-surfaceMesh_sphere": "surf/lh.sphere",
+            "pipeline-fs_hemi-R_desc-surfaceMesh_sphere": "surf/rh.sphere",
+            "pipeline-fs_hemi-L_desc-surfaceMap_sulc": "surf/lh.sulc",
+            "pipeline-fs_hemi-R_desc-surfaceMap_sulc": "surf/rh.sulc",
+            "pipeline-fs_hemi-L_desc-surfaceMap_thickness": "surf/lh.thickness",
+            "pipeline-fs_hemi-R_desc-surfaceMap_thickness": "surf/rh.thickness",
+            "pipeline-fs_hemi-L_desc-surfaceMap_volume": "surf/lh.volume",
+            "pipeline-fs_hemi-R_desc-surfaceMap_volume": "surf/rh.volume",
+            "pipeline-fs_hemi-L_desc-surfaceMesh_white": "surf/lh.white",
+            "pipeline-fs_hemi-R_desc-surfaceMesh_white": "surf/rh.white",
+            "pipeline-fs_xfm": "mri/transforms/talairach.lta",
+        }
+
+        for key, outfile in recon_outs.items():
+            fullpath = os.path.join(fs_path, outfile)
+            if os.path.exists(fullpath):
+                fs_ingress = create_general_datasource(f"gather_fs_{key}_dir")
+                fs_ingress.inputs.inputnode.set(
+                    unique_id=self.unique_id,
+                    data=fullpath,
+                    creds_path=self.creds_path,
+                    dl_dir=self.cfg.pipeline_setup["working_directory"]["path"],
+                )
+                self.set_data(
+                    key, fs_ingress, "outputspec.data", {}, "", f"fs_{key}_ingress"
+                )
+            else:
+                warnings.warn(
+                    str(LookupError(f"\n[!] Path does not exist for {fullpath}.\n"))
+                )
+
+        return
+
+    def ingress_output_dir(self) -> None:
+        """Ingress an output directory into a ResourcePool."""
+        dir_path = self.data_paths["derivatives_dir"]
+
+        WFLOGGER.info("\nPulling outputs from %s.\n", dir_path)
+
+        anat = os.path.join(dir_path, "anat")
+        func = os.path.join(dir_path, "func")
+
+        outdir_anat = []
+        outdir_func = []
+        func_paths = {}
+        func_dict = {}
+        func_key = ""
+
+        for subdir in [anat, func]:
+            if os.path.isdir(subdir):
+                for filename in os.listdir(subdir):
+                    for ext in EXTS:
+                        if ext in filename:
+                            if subdir == anat:
+                                outdir_anat.append(os.path.join(subdir, filename))
+                            else:
+                                outdir_func.append(os.path.join(subdir, filename))
+
+        # Add derivatives directory to rpool
+        ingress = create_general_datasource("gather_derivatives_dir")
+        ingress.inputs.inputnode.set(
+            unique_id=self.unique_id,
+            data=dir_path,
+            creds_path=self.creds_path,
+            dl_dir=self.cfg.pipeline_setup["working_directory"]["path"],
+        )
+        self.set_data(
+            "derivatives-dir",
+            ingress,
+            "outputspec.data",
+            {},
+            "",
+            "outdir_config_ingress",
+        )
+
+        for subdir in [outdir_anat, outdir_func]:
+            for filepath in subdir:
+                filename = str(filepath)
+                for ext in EXTS:
+                    filename = filename.split("/")[-1].replace(ext, "")
+
+                data_label = filename.split(self.unique_id)[1].lstrip("_")
+
+                if len(filename) == len(data_label):
+                    msg = (
+                        "\n\n[!] Possibly wrong participant or "
+                        "session in this directory?\n\n"
+                        f"Filepath: {filepath}\n\n"
+                    )
+                    raise Exception(msg)
+
+                bidstag = ""
+                for tag in data_label.split("_"):
+                    for prefix in ["task-", "run-", "acq-", "rec"]:
+                        if tag.startswith(prefix):
+                            bidstag += f"{tag}_"
+                            data_label = data_label.replace(f"{tag}_", "")
+                data_label, json = strip_template(data_label)
+
+                json_info, pipe_idx, node_name, data_label = json_outdir_ingress(
+                    filepath, data_label, json
+                )
+
+                if (
+                    "template" in data_label
+                    and not json_info["Template"]
+                    == self.cfg.pipeline_setup["outdir_ingress"]["Template"]
+                ):
+                    continue
+                # Rename confounds to avoid confusion in nuisance regression
+                if data_label.endswith("desc-confounds_timeseries"):
+                    data_label = "pipeline-ingress_desc-confounds_timeseries"
+
+                if len(bidstag) > 1:
+                    # Remove tail symbol
+                    bidstag = bidstag[:-1]
+                    if bidstag.startswith("task-"):
+                        bidstag = bidstag.replace("task-", "")
+
+                # Rename bold mask for CPAC naming convention
+                # and to avoid collision with anat brain mask
+                if data_label.endswith("desc-brain_mask") and filepath in outdir_func:
+                    data_label = data_label.replace("brain_mask", "bold_mask")
+
+                try:
+                    pipe_x = self.get_pipe_number(pipe_idx)
+                except ValueError:
+                    pipe_x = len(self.pipe_list)
+                if filepath in outdir_anat:
+                    ingress = create_general_datasource(
+                        f"gather_anat_outdir_{data_label!s}_{pipe_x}"
+                    )
+                    ingress.inputs.inputnode.set(
+                        unique_id=self.unique_id,
+                        data=filepath,
+                        creds_path=self.creds_path,
+                        dl_dir=self.cfg.pipeline_setup["working_directory"]["path"],
+                    )
+                    self.set_data(
+                        data_label,
+                        ingress,
+                        "outputspec.data",
+                        json_info,
+                        pipe_idx,
+                        node_name,
+                        f"outdir_{data_label}_ingress",
+                        inject=True,
+                    )
+                else:
+                    if data_label.endswith("desc-preproc_bold"):
+                        func_key = data_label
+                        func_dict[bidstag] = {}
+                        func_dict[bidstag]["scan"] = str(filepath)
+                        func_dict[bidstag]["scan_parameters"] = json_info
+                        func_dict[bidstag]["pipe_idx"] = pipe_idx
+                    if data_label.endswith("desc-brain_mask"):
+                        data_label = data_label.replace("brain_mask", "bold_mask")
+                    try:
+                        func_paths[data_label].append(filepath)
+                    except (AttributeError, KeyError, TypeError):
+                        func_paths[data_label] = []
+                        func_paths[data_label].append(filepath)
+
+        if func_dict:
+            self.func_outdir_ingress(
+                func_dict,
+                func_key,
+                func_paths,
+            )
+
+        if self.cfg.surface_analysis["freesurfer"]["ingress_reconall"]:
+            self.ingress_freesurfer()
+
+    def ingress_func_metadata(
+        self,
+        num_strat=None,
+    ) -> tuple[bool, bool, list[str]]:
+        """Ingress metadata for functional scans."""
+        name_suffix = ""
+        for suffix_part in (self.unique_id, num_strat):
+            if suffix_part is not None:
+                name_suffix += f"_{suffix_part}"
+        # Grab field maps
+        diff = False
+        blip = False
+        fmap_rp_list = []
+        fmap_TE_list = []
+        if "fmap" in self.data_paths:
+            second = False
+            for orig_key in self.data_paths["fmap"]:
+                gather_fmap = create_fmap_datasource(
+                    self.data_paths["fmap"], f"fmap_gather_{orig_key}_{self.part_id}"
+                )
+                gather_fmap.inputs.inputnode.set(
+                    subject=self.part_id,
+                    creds_path=self.creds_path,
+                    dl_dir=self.cfg.pipeline_setup["working_directory"]["path"],
+                )
+                gather_fmap.inputs.inputnode.scan = orig_key
+
+                key = orig_key
+                if "epi" in key and not second:
+                    key = "epi-1"
+                    second = True
+                elif "epi" in key and second:
+                    key = "epi-2"
+
+                self.set_data(
+                    key, gather_fmap, "outputspec.rest", {}, "", "fmap_ingress"
+                )
+                self.set_data(
+                    f"{key}-scan-params",
+                    gather_fmap,
+                    "outputspec.scan_params",
+                    {},
+                    "",
+                    "fmap_params_ingress",
+                )
+
+                fmap_rp_list.append(key)
+
+                get_fmap_metadata_imports = ["import json"]
+                get_fmap_metadata = pe.Node(
+                    Function(
+                        input_names=["data_config_scan_params"],
+                        output_names=[
+                            "dwell_time",
+                            "pe_direction",
+                            "total_readout",
+                            "echo_time",
+                            "echo_time_one",
+                            "echo_time_two",
+                        ],
+                        function=get_fmap_phasediff_metadata,
+                        imports=get_fmap_metadata_imports,
+                    ),
+                    name=f"{key}_get_metadata{name_suffix}",
+                )
+
+                self._init_wf.connect(
+                    gather_fmap,
+                    "outputspec.scan_params",
+                    get_fmap_metadata,
+                    "data_config_scan_params",
+                )
+
+                if "phase" in key:
+                    # leave it open to all three options, in case there is a
+                    # phasediff image with either a single EchoTime field (which
+                    # usually matches one of the magnitude EchoTimes), OR
+                    # a phasediff with an EchoTime1 and EchoTime2
+
+                    # at least one of these rpool keys will have a None value,
+                    # which will be sorted out in gather_echo_times below
+                    self.set_data(
+                        f"{key}-TE",
+                        get_fmap_metadata,
+                        "echo_time",
+                        {},
+                        "",
+                        "fmap_TE_ingress",
+                    )
+                    fmap_TE_list.append(f"{key}-TE")
+
+                    self.set_data(
+                        f"{key}-TE1",
+                        get_fmap_metadata,
+                        "echo_time_one",
+                        {},
+                        "",
+                        "fmap_TE1_ingress",
+                    )
+                    fmap_TE_list.append(f"{key}-TE1")
+
+                    self.set_data(
+                        f"{key}-TE2",
+                        get_fmap_metadata,
+                        "echo_time_two",
+                        {},
+                        "",
+                        "fmap_TE2_ingress",
+                    )
+                    fmap_TE_list.append(f"{key}-TE2")
+
+                elif "magnitude" in key:
+                    self.set_data(
+                        f"{key}-TE",
+                        get_fmap_metadata,
+                        "echo_time",
+                        {},
+                        "",
+                        "fmap_TE_ingress",
+                    )
+                    fmap_TE_list.append(f"{key}-TE")
+
+                self.set_data(
+                    f"{key}-dwell",
+                    get_fmap_metadata,
+                    "dwell_time",
+                    {},
+                    "",
+                    "fmap_dwell_ingress",
+                )
+                self.set_data(
+                    f"{key}-pedir",
+                    get_fmap_metadata,
+                    "pe_direction",
+                    {},
+                    "",
+                    "fmap_pedir_ingress",
+                )
+                self.set_data(
+                    f"{key}-total-readout",
+                    get_fmap_metadata,
+                    "total_readout",
+                    {},
+                    "",
+                    "fmap_readout_ingress",
+                )
+
+                if "phase" in key or "mag" in key:
+                    diff = True
+
+                if re.match("epi_[AP]{2}", orig_key):
+                    blip = True
+
+            if diff:
+                calc_delta_ratio = pe.Node(
+                    Function(
+                        input_names=["effective_echo_spacing", "echo_times"],
+                        output_names=["deltaTE", "ees_asym_ratio"],
+                        function=calc_delta_te_and_asym_ratio,
+                        imports=["from typing import Optional"],
+                    ),
+                    name=f"diff_distcor_calc_delta{name_suffix}",
+                )
+
+                gather_echoes = pe.Node(
+                    Function(
+                        input_names=[
+                            "echotime_1",
+                            "echotime_2",
+                            "echotime_3",
+                            "echotime_4",
+                        ],
+                        output_names=["echotime_list"],
+                        function=gather_echo_times,
+                    ),
+                    name="fugue_gather_echo_times",
+                )
+
+                for idx, fmap_file in enumerate(fmap_TE_list, start=1):
+                    try:
+                        node, out_file = self.get(fmap_file)[
+                            f"['{fmap_file}:fmap_TE_ingress']"
+                        ]["data"]
+                        self._init_wf.connect(
+                            node, out_file, gather_echoes, f"echotime_{idx}"
+                        )
+                    except KeyError:
+                        pass
+
+                self._init_wf.connect(
+                    gather_echoes, "echotime_list", calc_delta_ratio, "echo_times"
+                )
+
+        # Add in nodes to get parameters from configuration file
+        # a node which checks if scan_parameters are present for each scan
+        scan_params = pe.Node(
+            Function(
+                input_names=[
+                    "data_config_scan_params",
+                    "subject_id",
+                    "scan",
+                    "pipeconfig_tr",
+                    "pipeconfig_tpattern",
+                    "pipeconfig_start_indx",
+                    "pipeconfig_stop_indx",
+                ],
+                output_names=[
+                    "tr",
+                    "tpattern",
+                    "template",
+                    "ref_slice",
+                    "start_indx",
+                    "stop_indx",
+                    "pe_direction",
+                    "effective_echo_spacing",
+                ],
+                function=get_scan_params,
+                imports=["from CPAC.utils.utils import check, try_fetch_parameter"],
+            ),
+            name=f"bold_scan_params_{self.part_id}{name_suffix}",
+        )
+        scan_params.inputs.subject_id = self.part_id
+        scan_params.inputs.set(
+            pipeconfig_start_indx=self.cfg.functional_preproc["truncation"]["start_tr"],
+            pipeconfig_stop_indx=self.cfg.functional_preproc["truncation"]["stop_tr"],
+        )
+
+        node, out = self.get("scan")["['scan:func_ingress']"]["data"]
+        self._init_wf.connect(node, out, scan_params, "scan")
+
+        # Workaround for extracting metadata with ingress
+        if self.check_rpool("derivatives-dir"):
+            selectrest_json = pe.Node(
+                Function(
+                    input_names=["scan", "rest_dict", "resource"],
+                    output_names=["file_path"],
+                    function=get_rest,
+                    as_module=True,
+                ),
+                name="selectrest_json",
+            )
+            selectrest_json.inputs.rest_dict = self.data_paths
+            selectrest_json.inputs.resource = "scan_parameters"
+            self._init_wf.connect(node, out, selectrest_json, "scan")
+            self._init_wf.connect(
+                selectrest_json, "file_path", scan_params, "data_config_scan_params"
+            )
+
+        else:
+            # wire in the scan parameter workflow
+            node, out = self.get("scan-params")["['scan-params:scan_params_ingress']"][
+                "data"
+            ]
+            self._init_wf.connect(node, out, scan_params, "data_config_scan_params")
+
+        self.set_data("TR", scan_params, "tr", {}, "", "func_metadata_ingress")
+        self.set_data(
+            "tpattern", scan_params, "tpattern", {}, "", "func_metadata_ingress"
+        )
+        self.set_data(
+            "template", scan_params, "template", {}, "", "func_metadata_ingress"
+        )
+        self.set_data(
+            "start-tr", scan_params, "start_indx", {}, "", "func_metadata_ingress"
+        )
+        self.set_data(
+            "stop-tr", scan_params, "stop_indx", {}, "", "func_metadata_ingress"
+        )
+        self.set_data(
+            "pe-direction", scan_params, "pe_direction", {}, "", "func_metadata_ingress"
+        )
+
+        if diff:
+            # Connect EffectiveEchoSpacing from functional metadata
+            self.set_data(
+                "effectiveEchoSpacing",
+                scan_params,
+                "effective_echo_spacing",
+                {},
+                "",
+                "func_metadata_ingress",
+            )
+            node, out_file = self.get("effectiveEchoSpacing")[
+                "['effectiveEchoSpacing:func_metadata_ingress']"
+            ]["data"]
+            self._init_wf.connect(
+                node, out_file, calc_delta_ratio, "effective_echo_spacing"
+            )
+            self.set_data(
+                "deltaTE", calc_delta_ratio, "deltaTE", {}, "", "deltaTE_ingress"
+            )
+            self.set_data(
+                "ees-asym-ratio",
+                calc_delta_ratio,
+                "ees_asym_ratio",
+                {},
+                "",
+                "ees_asym_ratio_ingress",
+            )
+
+        return diff, blip, fmap_rp_list
+
+    def ingress_pipeconfig_paths(self):
+        """Ingress config file paths."""
+        # TODO: may want to change the resource keys for each to include one level up in the YAML as well
+
+        import pandas as pd
+        import pkg_resources as p
+
+        template_csv = p.resource_filename("CPAC", "resources/cpac_templates.csv")
+        template_df = pd.read_csv(template_csv, keep_default_na=False)
+
+        for row in template_df.itertuples():
+            key = row.Key
+            val = row.Pipeline_Config_Entry
+            val = self.cfg.get_nested(self.cfg, [x.lstrip() for x in val.split(",")])
+            resolution = row.Intended_Resolution_Config_Entry
+            desc = row.Description
+
+            if not val:
+                continue
+
+            if resolution:
+                res_keys = [x.lstrip() for x in resolution.split(",")]
+                tag = res_keys[-1]
+            json_info = {}
+
+            if "$FSLDIR" in val:
+                val = val.replace(
+                    "$FSLDIR", self.cfg.pipeline_setup["system_config"]["FSLDIR"]
+                )
+            if "$priors_path" in val:
+                priors_path = (
+                    self.cfg.segmentation["tissue_segmentation"]["FSL-FAST"][
+                        "use_priors"
+                    ]["priors_path"]
+                    or ""
+                )
+                if "$FSLDIR" in priors_path:
+                    priors_path = priors_path.replace(
+                        "$FSLDIR", self.cfg.pipeline_setup["system_config"]["FSLDIR"]
+                    )
+                val = val.replace("$priors_path", priors_path)
+            if "${resolution_for_anat}" in val:
+                val = val.replace(
+                    "${resolution_for_anat}",
+                    self.cfg.registration_workflows["anatomical_registration"][
+                        "resolution_for_anat"
+                    ],
+                )
+            if "${func_resolution}" in val:
+                val = val.replace(
+                    "${func_resolution}",
+                    self.cfg.registration_workflows["functional_registration"][
+                        "func_registration_to_template"
+                    ]["output_resolution"][tag],
+                )
+
+            if desc:
+                template_name, _template_desc = lookup_identifier(val)
+                if template_name:
+                    desc = f"{template_name} - {desc}"
+                json_info["Description"] = f"{desc} - {val}"
+            if resolution:
+                resolution = self.cfg.get_nested(self.cfg, res_keys)
+                json_info["Resolution"] = resolution
+
+                resampled_template = pe.Node(
+                    Function(
+                        input_names=["resolution", "template", "template_name", "tag"],
+                        output_names=["resampled_template"],
+                        function=resolve_resolution,
+                        as_module=True,
+                    ),
+                    name="resampled_" + key,
+                )
+
+                resampled_template.inputs.resolution = resolution
+                resampled_template.inputs.template = val
+                resampled_template.inputs.template_name = key
+                resampled_template.inputs.tag = tag
+
+                # the set_data below is set up a little differently, because we are
+                # injecting and also over-writing already-existing entries
+                # other alternative would have been to ingress into the
+                # resampled_template node from the already existing entries, but we
+                # didn't do that here
+                self.set_data(
+                    key,
+                    resampled_template,
+                    "resampled_template",
+                    json_info,
+                    "",
+                    "template_resample",
+                )  # pipe_idx (after the blank json {}) should be the previous strat that you want deleted! because you're not connecting this the regular way, you have to do it manually
+
+            elif val:
+                config_ingress = create_general_datasource(f"gather_{key}")
+                config_ingress.inputs.inputnode.set(
+                    unique_id=self.unique_id,
+                    data=val,
+                    creds_path=self.creds_path,
+                    dl_dir=self.cfg.pipeline_setup["working_directory"]["path"],
+                )
+                self.set_data(
+                    key,
+                    config_ingress,
+                    "outputspec.data",
+                    json_info,
+                    "",
+                    f"{key}_config_ingress",
+                )
+
+    def ingress_raw_func_data(self):
+        """Ingress raw functional data."""
+        func_paths_dct = self.data_paths["func"]
+
+        func_wf = self.create_func_datasource(
+            func_paths_dct, f"func_ingress_{self.part_id}_{self.ses_id}"
+        )
+        func_wf.inputs.inputnode.set(
+            subject=self.part_id,
+            creds_path=self.creds_path,
+            dl_dir=self.cfg.pipeline_setup["working_directory"]["path"],
+        )
+        func_wf.get_node("inputnode").iterables = ("scan", list(func_paths_dct.keys()))
+
+        self.set_data("subject", func_wf, "outputspec.subject", {}, "", "func_ingress")
+        self.set_data("bold", func_wf, "outputspec.rest", {}, "", "func_ingress")
+        self.set_data("scan", func_wf, "outputspec.scan", {}, "", "func_ingress")
+        self.set_data(
+            "scan-params",
+            func_wf,
+            "outputspec.scan_params",
+            {},
+            "",
+            "scan_params_ingress",
+        )
+
+        # TODO: CHECK FOR PARAMETERS
+
+        diff, blip, fmap_rp_list = self.ingress_func_metadata()
+
+        # Memoize list of local functional scans
+        # TODO: handle S3 files
+        # Skip S3 files for now
+
+        local_func_scans = [
+            func_paths_dct[scan]["scan"]
+            for scan in func_paths_dct.keys()
+            if not func_paths_dct[scan]["scan"].startswith("s3://")
+        ]
+        if local_func_scans:
+            # pylint: disable=protected-access
+            self._init_wf._local_func_scans = local_func_scans
+            if self.cfg.pipeline_setup["Debugging"]["verbose"]:
+                verbose_logger = getLogger("CPAC.engine")
+                verbose_logger.debug("local_func_scans: %s", local_func_scans)
+        del local_func_scans
+
+        return diff, blip, fmap_rp_list
+
+    def func_outdir_ingress(self, func_dict: dict, key: str, func_paths: dict) -> None:
+        """Ingress a functional output directory."""
+        pipe_x = len(self.pipe_list)
+        ingress = self.create_func_datasource(
+            func_dict, f"gather_func_outdir_{key}_{pipe_x}"
+        )
+        ingress.inputs.inputnode.set(
+            subject=self.unique_id,
+            creds_path=self.creds_path,
+            dl_dir=self.cfg.pipeline_setup["working_directory"]["path"],
+        )
+        self.set_data("subject", ingress, "outputspec.subject", {}, "", "func_ingress")
+        ingress.get_node("inputnode").iterables = ("scan", list(func_dict.keys()))
+        self.set_data(key, ingress, "outputspec.rest", {}, "", "func_ingress")
+
+        self.set_data("scan", ingress, "outputspec.scan", {}, "", "func_ingress")
+        self.set_data(
+            "scan-params",
+            ingress,
+            "outputspec.scan_params",
+            {},
+            "",
+            "scan_params_ingress",
+        )
+        self.ingress_func_metadata()
+
+        # Have to do it this weird way to save the parsed BIDS tag & filepath
+        mask_paths_key = (
+            "desc-bold_mask"
+            if "desc-bold_mask" in func_paths
+            else "space-template_desc-bold_mask"
+        )
+        ts_paths_key = "pipeline-ingress_desc-confounds_timeseries"
+
+        # Connect func data with approproate scan name
+        iterables = pe.Node(
+            Function(
+                input_names=["scan", "mask_paths", "ts_paths"],
+                output_names=["out_scan", "mask", "confounds"],
+                function=set_iterables,
+            ),
+            name=f"set_iterables_{pipe_x}",
+        )
+        iterables.inputs.mask_paths = func_paths[mask_paths_key]
+        iterables.inputs.ts_paths = func_paths[ts_paths_key]
+        self._init_wf.connect(ingress, "outputspec.scan", iterables, "scan")
+
+        for key in func_paths:
+            if key in (mask_paths_key, ts_paths_key):
+                ingress_func = create_general_datasource(f"ingress_func_data_{key}")
+                ingress_func.inputs.inputnode.set(
+                    unique_id=self.unique_id,
+                    creds_path=self.creds_path,
+                    dl_dir=self.cfg.pipeline_setup["working_directory"]["path"],
+                )
+                self._init_wf.connect(
+                    iterables, "out_scan", ingress_func, "inputnode.scan"
+                )
+                if key == mask_paths_key:
+                    self._init_wf.connect(
+                        iterables, "mask", ingress_func, "inputnode.data"
+                    )
+                    self.set_data(
+                        key,
+                        ingress_func,
+                        "inputnode.data",
+                        {},
+                        "",
+                        f"outdir_{key}_ingress",
+                    )
+                elif key == ts_paths_key:
+                    self._init_wf.connect(
+                        iterables, "confounds", ingress_func, "inputnode.data"
+                    )
+                    self.set_data(
+                        key,
+                        ingress_func,
+                        "inputnode.data",
+                        {},
+                        "",
+                        f"outdir_{key}_ingress",
+                    )
+
+    def ingress_raw_anat_data(self) -> None:
+        """Ingress raw anatomical data."""
+        if "anat" not in self.data_paths:
+            WFLOGGER.warning("No anatomical data present.")
+            return
+
+        anat_flow = create_anat_datasource(
+            f"anat_T1w_gather_{self.part_id}_{self.ses_id}"
+        )
+
+        anat = {}
+        if isinstance(self.data_paths["anat"], str):
+            anat["T1"] = self.data_paths["anat"]
+        elif "T1w" in self.data_paths["anat"]:
+            anat["T1"] = self.data_paths["anat"]["T1w"]
+
+        if "T1" in anat:
+            anat_flow.inputs.inputnode.set(
+                subject=self.part_id,
+                anat=anat["T1"],
+                creds_path=self.creds_path,
+                dl_dir=self.cfg.pipeline_setup["working_directory"]["path"],
+                img_type="anat",
+            )
+            self.set_data("T1w", anat_flow, "outputspec.anat", {}, "", "anat_ingress")
+
+        if "T2w" in self.data_paths["anat"]:
+            anat_flow_T2 = create_anat_datasource(
+                f"anat_T2w_gather_{self.part_id}_{self.ses_id}"
+            )
+            anat_flow_T2.inputs.inputnode.set(
+                subject=self.part_id,
+                anat=self.data_paths["anat"]["T2w"],
+                creds_path=self.creds_path,
+                dl_dir=self.cfg.pipeline_setup["working_directory"]["path"],
+                img_type="anat",
+            )
+            self.set_data(
+                "T2w", anat_flow_T2, "outputspec.anat", {}, "", "anat_ingress"
+            )
+
+        if self.cfg.surface_analysis["freesurfer"]["ingress_reconall"]:
+            self.ingress_freesurfer()
+
+
+def initiate_rpool(
+    wf: pe.Workflow,
+    cfg: Configuration,
+    data_paths: Optional[dict] = None,
+    part_id: Optional[str] = None,
+) -> ResourcePool:
     """
     Initialize a new ResourcePool.
 
-    data_paths format:
+    data_paths format::
+
       {'anat': {
             'T1w': '{T1w path}',
             'T2w': '{T2w path}'
@@ -1480,20 +2579,32 @@ def initiate_rpool(wf, cfg, data_paths=None, part_id=None):
     # TODO: used for BIDS-Derivatives (below), and possible refactoring of
     # TODO: the raw data config to use 'T1w' label instead of 'anat' etc.
 
+    kwargs = {"cfg": cfg, "wf": wf}
     if data_paths:
-        part_id = data_paths["subject_id"]
-        ses_id = data_paths["unique_id"]
+        part_id: str = data_paths["subject_id"]
+        ses_id: str = data_paths["unique_id"]
         if "creds_path" not in data_paths:
             creds_path = None
         else:
-            creds_path = data_paths["creds_path"]
-        unique_id = f"{part_id}_{ses_id}"
-
+            creds_path: Optional[Path | str] = data_paths["creds_path"]
+        unique_id: str = f"{part_id}_{ses_id}"
+        kwargs.update(
+            {
+                "part_id": part_id,
+                "ses_id": ses_id,
+                "creds_path": creds_path,
+                "data_paths": data_paths,
+            }
+        )
     elif part_id:
         unique_id = part_id
         creds_path = None
+        kwargs.update({"part_id": part_id, "creds_path": creds_path})
+    else:
+        unique_id = ""
+    kwargs.update({"unique_id": unique_id})
 
-    rpool = ResourcePool(name=unique_id, cfg=cfg)
+    rpool = ResourcePool(name=unique_id, **kwargs)
 
     if data_paths:
         # ingress outdir
@@ -1502,32 +2613,25 @@ def initiate_rpool(wf, cfg, data_paths=None, part_id=None):
                 data_paths["derivatives_dir"]
                 and cfg.pipeline_setup["outdir_ingress"]["run"]
             ):
-                wf, rpool = ingress_output_dir(
-                    wf,
-                    cfg,
-                    rpool,
-                    unique_id,
-                    data_paths,
-                    part_id,
-                    ses_id,
-                    creds_path=None,
-                )
-        except:
-            rpool = ingress_raw_anat_data(
-                wf, rpool, cfg, data_paths, unique_id, part_id, ses_id
-            )
+                rpool.ingress_output_dir()
+        except (AttributeError, KeyError):
+            rpool.ingress_raw_anat_data()
             if "func" in data_paths:
-                wf, rpool, diff, blip, fmap_rp_list = ingress_raw_func_data(
-                    wf, rpool, cfg, data_paths, unique_id, part_id, ses_id
-                )
+                rpool.ingress_raw_func_data()
 
     # grab any file paths from the pipeline config YAML
-    rpool = ingress_pipeconfig_paths(cfg, rpool, unique_id, creds_path)
+    rpool.ingress_pipeconfig_paths()
 
     # output files with 4 different scans
 
-    return (wf, rpool)
+    return rpool._init_wf, rpool
 
 
 class StratPool(ResourcePool):
     """All resources for a strategy."""
+
+    def __init__(self):
+        """Initialize a ResourcePool."""
+
+    def append_name(self, name):
+        self.name.append(name)
